Add --ide flag to verify-config, changing some errors into warnings when run from the IDE.
