+++ mirrord/cli/src/extension.rs
@@ -61,7 +61,7 @@ pub(crate) async fn extension_exec(args: ExtensionExecArgs, watch: drain::Watch)
 
     let mut analytics = AnalyticsReporter::only_error(config.telemetry, watch);
 
-    config.verify(&mut context)?;
+    config.verify(&mut context, false)?;
     for warning in context.get_warnings() {
         progress.warning(warning);
     }
