+++ mirrord/config/src/lib.rs
@@ -323,7 +323,11 @@ impl LayerConfig {
     /// We don't call it from `from_env` since we want to verify it only once (from cli)
     ///
     /// Fills `context` with the warnings.
-    pub fn verify(&self, context: &mut ConfigContext) -> Result<(), ConfigError> {
+    ///
+    /// - `ide`: Identifies if this is being called from an IDE context, when using
+    /// `mirrord verify-config`. Turns some _target missing_ errors into warnings, as the target can
+    /// be selected after `verify-config` is run.
+    pub fn verify(&self, context: &mut ConfigContext, ide: bool) -> Result<(), ConfigError> {
         if self.pause {
             if self.agent.ephemeral && !self.agent.privileged {
                 context.add_warning("The target pause feature with ephemeral requires to enable the privileged flag on the agent.".to_string());
@@ -357,8 +361,9 @@ impl LayerConfig {
         {
             Err(ConfigError::Conflict(
                 "Cannot use both HTTP header filter and path filter at the same time".to_string(),
-            ))?;
+            ))?
         }
+
         if self
             .feature
             .network
@@ -381,24 +386,67 @@ impl LayerConfig {
                     .header_filter
                     .is_some())
         {
-            Err(ConfigError::Conflict("Cannot use old http filter and new http filter at the same time. Use only `http_filter` instead of `http_header_filter`".to_string()))?;
+            Err(ConfigError::Conflict("Cannot use old http filter and new http filter at the same time. Use only `http_filter` instead of `http_header_filter`".to_string()))?
         }
+
         if self.target.path.is_none() {
-            if self.target.namespace.is_some() {
-                Err(ConfigError::TargetNamespaceWithoutTarget)?;
-            }
-            if self.feature.network.incoming.is_steal() {
-                context.add_warning("Steal mode is not compatible with a targetless agent, please either disable this option or specify a target.".into());
-            }
-            if self.agent.ephemeral {
-                Err(ConfigError::Conflict(
-                    "Using an ephemeral container for the agent is not compatible with a targetless agent, please either disable this option or specify a target.".into(),
-                ))?;
-            }
-            if self.pause {
-                Err(ConfigError::Conflict(
-                    "The target pause feature is not compatible with a targetless agent, please either disable this option or specify a target.".into(),
-                ))?;
+            // In the IDE, a target may be selected after `mirrord verify-config` is run, so we
+            // for this case we treat these as warnings. They'll become errors once mirrord proper
+            // tries to start (if the user somehow managed to not select a target by then).
+            if ide {
+                if self.target.namespace.is_some() {
+                    context.add_warning(
+                        "A target namespace was specified, but no target was specified. \
+                            The IDE will prompt you to select a target."
+                            .into(),
+                    );
+                }
+
+                if self.feature.network.incoming.is_steal() {
+                    context.add_warning("Steal mode is not compatible with a targetless agent, please either disable this option or specify a target.".into());
+                }
+
+                if self.agent.ephemeral {
+                    context.add_warning(
+                        "Using an ephemeral container for the agent is not \
+                        compatible with a targetless agent, please either disable this option or \
+                        specify a target."
+                            .into(),
+                    );
+                }
+
+                if self.pause {
+                    context.add_warning(
+                        "The target pause feature is not compatible with a targetless agent, please \
+                        either disable this option or specify a target."
+                            .into()
+                    );
+                }
+            } else {
+                if self.target.namespace.is_some() {
+                    Err(ConfigError::TargetNamespaceWithoutTarget)?
+                }
+
+                if self.feature.network.incoming.is_steal() {
+                    Err(ConfigError::Conflict("Steal mode is not compatible with a targetless agent, please either disable this option or specify a target.".into()))?
+                }
+
+                if self.agent.ephemeral {
+                    Err(ConfigError::Conflict(
+                        "Using an ephemeral container for the agent is not \
+                         compatible with a targetless agent, please either disable this option or \
+                        specify a target."
+                            .into(),
+                    ))?
+                }
+
+                if self.pause {
+                    Err(ConfigError::Conflict(
+                        "The target pause feature is not compatible with a \
+                        targetless agent, please either disable this option or specify a target."
+                            .into(),
+                    ))?
+                }
             }
         }
         Ok(())
